bg_editor/requirements.txt
certifi==2022.12.7  
charset-normalizer==3.0.1  
idna==2.10  
Pillow==10.1.0  
requests==2.31.0  
urllib3==1.26.14
googletrans==4.0.0-rc1
qt-material==2.14
pyinstaller
jupyter
ipykernel